refactor: extract out typegen templating
